@@ -1,20 +1,13 @@
 package com.example;
 
-import com.example.infraestructure.database.person.PersonRepository;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.example.repository.impl.PersonRepositoryImpl;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import javax.sql.DataSource;
-
 import java.sql.SQLException;
 
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class DemoApplicationTests {
 
@@ -22,7 +15,7 @@ public class DemoApplicationTests {
     private DataSource dataSource;
 
     @Autowired
-    private PersonRepository repository;
+    private PersonRepositoryImpl repository;
 
     @Test
     public void test() {
